requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
python-dotenv
redis
rq
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
playwright
